requirements.txt
# this file is for pip, run it with 
#
# pip install -r requirements.txt
#

Django >= 1.4
mysql-python
django-tastypie

